+++ include/Gaffer/SplinePlug.h
@@ -142,7 +142,7 @@ class GAFFER_API SplinePlug : public ValuePlug
 		const T &defaultValue() const;
 		void setToDefault() override;
 		bool isSetToDefault() const override;
-
+		void resetDefault() override;
 
 		/// Sets the value of the points and interpolation child plugs
 		/// \undoable