+++ src/GafferBindings/ValuePlugBinding.cpp
@@ -46,6 +46,7 @@
 #include "Gaffer/Metadata.h"
 #include "Gaffer/Node.h"
 #include "Gaffer/Reference.h"
+#include "Gaffer/Spreadsheet.h"
 #include "Gaffer/ValuePlug.h"
 
 #include "boost/algorithm/string/predicate.hpp"
@@ -284,12 +285,32 @@ std::string ValuePlugSerialiser::postHierarchy( const Gaffer::GraphComponent *gr
 	{
 		// Top level ValuePlug. We are responsible for emitting the
 		// appropriate `setValue()` calls for this and all descendants.
-		if( !shouldResetPlugDefault( plug, &serialisation ) )
+
+		const bool resetDefault = shouldResetPlugDefault( plug, &serialisation );
+		const bool isRowsPlug = IECore::runTimeCast<const Spreadsheet::RowsPlug>( plug );
+
+		if( !resetDefault || isRowsPlug )
 		{
 			bool unused;
 			result = valueSerialisationWalk( plug, identifier, serialisation, unused ) + result;
 		}
 
+		if( isRowsPlug && resetDefault )
+		{
+			// RowsPlugs serialise their children independently via `addRow()` and `addColumn()`
+			// so we don't have the opportunity to reset default values in `repr()`. Instead we
+			// emit `setValue()` calls above and then call `resetDefault()` here afterwards.
+			//
+			/// \todo Replace the `shouldResetPlugDefault()` mechanism entirely. It only exists
+			/// to allow users to author default values when exporting references and extensions.
+			/// But this implicit introduction of new defaults during export is too error-prone.
+			/// It's very easy to change values for testing and then forget to change them back
+			/// before exporting again. Instead we should just expose `ValuePlug::resetDefault()`
+			/// to the user so that they can explicitly set the default at their convenience.
+			/// Then we should omit the `setValue()` calls when exporting instead, so that the
+			/// user can't inadvertently replace the default they carefully authored.
+			result += identifier + ".resetDefault()\n";
+		}
 	}
 
 	return result;